Added coverage

+++ src/utils/file_utilities.cc
@@ -26,9 +26,9 @@ void FileReader::unget() { m_file_ptr.unget(); }
 
 char FileReader::peek() { return m_file_ptr.peek(); }
 
-size_t FileReader::get_line() { return m_line; }
+int FileReader::get_line() { return m_line; }
 
-size_t FileReader::get_column() { return m_column; }
+int FileReader::get_column() { return m_column; }
 
 char FileReader::get_current_char() { return m_current_char; }
